First commit - single-threaded
This was a first attempt at trying to grab data on
the level of ambiguity of Wikipedia pages.

I quickly realized that this was going nowhere fast, since
this took over an hour to get through to "An" pages. This
commit is actually retrofitted, since I continued on with
the multithreading implementation (and added some other
stuff like logging, and end-case handling in PageProvider)
before realizing that this was going to take long enough
that I should make a Git repo.