Add multithreading
I don't have good data on how long this took to complete,
since I accidentally left the "Draining the Writing Queue"
step running long enough to fill up my terminal buffer (let
that be a lesson to me to *never* just write to `sysout`!),
but I did record that the fetching of data concluded about
40 minutes after I started the script - draining the
WritingQueue took a lot longer, however (several hours),
and I should probably find a way to speed that up.

This "worked", but I quickly found when analyzing the results
that there was some dirty data in there. Most notably, the
`<li>` elements in Table of Contents were being included, which
inflated complexity by up to 134.